--warning

+++ NSArray+YOLO.m
@@ -10,7 +10,7 @@
 - (NSArray *(^)(NSUInteger))chunk {
     return ^(NSUInteger size){
         id aa = [NSMutableArray new];
-        const int n = self.count / size;
+        const NSUInteger n = self.count / size;
         for (int x = 0; x < n; ++x)
             [aa addObject:self.slice(x*size, size)];
         return aa;
